exclude dev test articles from prior probabilities stats
