+++ bin/wiki_entity_linking/wiki_namespaces.py
@@ -1,0 +1,109 @@
+# coding: utf8
+from __future__ import unicode_literals
+
+
+# TODO: add more cases from non-English WP's
+
+# List of prefixes that refer to Wikipedia "file" pages
+FILE_NAMESPACE = [
+    "Bestand",
+    "File",
+]
+
+# List of prefixes that refer to Wikipedia "category" pages
+CATEGORY_NAMESPACE = [
+    "Kategori",
+    "Category",
+    "Categorie",
+]
+
+# List of prefixes that refer to Wikipedia "meta" pages
+# these will/should be matched ignoring case
+META_NAMESPACE = FILE_NAMESPACE + CATEGORY_NAMESPACE + [
+    "b",
+    "betawikiversity",
+    "Book",
+    "c",
+    "Commons",
+    "d",
+    "dbdump",
+    "download",
+    "Draft",
+    "Education",
+    "Foundation",
+    "Gadget",
+    "Gadget definition",
+    "Gebruiker",
+    "gerrit",
+    "Help",
+    "Image",
+    "Incubator",
+    "m",
+    "mail",
+    "mailarchive",
+    "media",
+    "MediaWiki",
+    "MediaWiki talk",
+    "Mediawikiwiki",
+    "MediaZilla",
+    "Meta",
+    "Metawikipedia",
+    "Module",
+    "mw",
+    "n",
+    "nost",
+    "oldwikisource",
+    "otrs",
+    "OTRSwiki",
+    "Overleg gebruiker",
+    "outreach",
+    "outreachwiki",
+    "Portal",
+    "phab",
+    "Phabricator",
+    "Project",
+    "q",
+    "quality",
+    "rev",
+    "s",
+    "spcom",
+    "Special",
+    "species",
+    "Strategy",
+    "sulutil",
+    "svn",
+    "Talk",
+    "Template",
+    "Template talk",
+    "Testwiki",
+    "ticket",
+    "TimedText",
+    "Toollabs",
+    "tools",
+    "tswiki",
+    "User",
+    "User talk",
+    "v",
+    "voy",
+    "w",
+    "Wikibooks",
+    "Wikidata",
+    "wikiHow",
+    "Wikinvest",
+    "wikilivres",
+    "Wikimedia",
+    "Wikinews",
+    "Wikipedia",
+    "Wikipedia talk",
+    "Wikiquote",
+    "Wikisource",
+    "Wikispecies",
+    "Wikitech",
+    "Wikiversity",
+    "Wikivoyage",
+    "wikt",
+    "wiktionary",
+    "wmf",
+    "wmania",
+    "WP",
+]
