+++ bin/wiki_entity_linking/wikidata_pretrain_kb.py
@@ -17,9 +17,8 @@ import logging
 from pathlib import Path
 import plac
 
-from bin.wiki_entity_linking import wikipedia_processor as wp, wikidata_processor as wd
+from bin.wiki_entity_linking import wikipedia_processor as wp, wikidata_processor as wd, wikipedia_processor
 from bin.wiki_entity_linking import kb_creator
-from bin.wiki_entity_linking import training_set_creator
 from bin.wiki_entity_linking import TRAINING_DATA_FILE, KB_FILE, ENTITY_DESCR_PATH, KB_MODEL_DIR, LOG_FORMAT
 from bin.wiki_entity_linking import ENTITY_FREQ_PATH, PRIOR_PROB_PATH, ENTITY_DEFS_PATH
 import spacy
@@ -40,7 +39,7 @@ logger = logging.getLogger(__name__)
     loc_prior_prob=("Location to file with prior probabilities", "option", "p", Path),
     loc_entity_defs=("Location to file with entity definitions", "option", "d", Path),
     loc_entity_desc=("Location to file with entity descriptions", "option", "s", Path),
-    descriptions_from_wikipedia=("Flag for using wp descriptions not wd", "flag", "wp"),
+    desc_from_wp=("Flag for using wp descriptions not wd", "flag", "wp"),
     limit_prior=("Threshold to limit lines read from WP for prior probabilities", "option", "lp", int),
     limit_train=("Threshold to limit lines read from WP for training set", "option", "lt", int),
     limit_wd=("Threshold to limit lines read from WD", "option", "lw", int),
@@ -58,7 +57,7 @@ def main(
     loc_prior_prob=None,
     loc_entity_defs=None,
     loc_entity_desc=None,
-    descriptions_from_wikipedia=False,
+    desc_from_wp=False,
     limit_prior=None,
     limit_train=None,
     limit_wd=None,
@@ -102,12 +101,12 @@ def main(
     # STEP 3: calculate entity frequencies
     if not entity_freq_path.exists():
         logger.info("STEP 3: Calculating and writing entity frequencies to {}".format(entity_freq_path))
-        wp.write_entity_counts(prior_prob_path, entity_freq_path, to_print=False)
+        wp.write_entity_counts(prior_prob_path, entity_freq_path)
     else:
         logger.info("STEP 3: Reading entity frequencies from {}".format(entity_freq_path))
 
     # STEP 4: reading definitions and (possibly) descriptions from WikiData or from file
-    if (not entity_defs_path.exists()) or (not descriptions_from_wikipedia and not entity_descr_path.exists()):
+    if (not entity_defs_path.exists()) or (not desc_from_wp and not entity_descr_path.exists()):
         # It takes about 10h to process 55M lines of Wikidata JSON dump
         logger.info("STEP 4: Parsing and writing Wikidata entity definitions to {}".format(entity_defs_path))
         if limit_wd is not None:
@@ -117,35 +116,35 @@ def main(
             limit_wd,
             to_print=False,
             lang=lang,
-            parse_descriptions=(not descriptions_from_wikipedia),
+            parse_descriptions=(not desc_from_wp),
         )
         wd.write_entity_files(entity_defs_path, title_to_id)
-        if not descriptions_from_wikipedia:
+        if not desc_from_wp:
             logger.info("STEP 4b: Writing Wikidata entity descriptions to {}".format(entity_descr_path))
             wd.write_entity_description_files(entity_descr_path, id_to_descr)
     else:
         logger.info("STEP 4: Reading entity definitions from {}".format(entity_defs_path))
-        if not descriptions_from_wikipedia:
+        if not desc_from_wp:
             logger.info("STEP 4b: Reading entity descriptions from {}".format(entity_descr_path))
 
     # STEP 5: Getting gold entities from Wikipedia
-    if (not training_entities_path.exists()) or (descriptions_from_wikipedia and not entity_descr_path.exists()):
+    if (not training_entities_path.exists()) or (desc_from_wp and not entity_descr_path.exists()):
         logger.info("STEP 5: Parsing and writing Wikipedia gold entities to {}".format(training_entities_path))
         if limit_train is not None:
             logger.warning("Warning: reading only {} lines of Wikidata dump".format(limit_train))
-        training_set_creator.create_training_examples_and_descriptions(
+        wikipedia_processor.create_training_and_desc(
             wp_xml,
             entity_defs_path,
             entity_descr_path,
             training_entities_path,
-            parse_descriptions=descriptions_from_wikipedia,
+            parse_desc=desc_from_wp,
             limit=limit_train,
         )
-        if descriptions_from_wikipedia:
+        if desc_from_wp:
             logger.info("STEP 5b: Parsing and writing Wikipedia descriptions to {}".format(entity_descr_path))
     else:
         logger.info("STEP 5: Reading gold entities from {}".format(training_entities_path))
-        if descriptions_from_wikipedia:
+        if desc_from_wp:
             logger.info("STEP 5b: Reading entity descriptions from {}".format(entity_descr_path))
 
     # STEP 6: creating the actual KB
